Improved new API (#142)
