+++ neural_compressor/benchmark.py
@@ -18,6 +18,8 @@
 from .utils import logger
 from .data import DATALOADERS
 from .experimental import Benchmark as ExpBenchmark
+from .conf.pythonic_config import Config
+from .config import BenchmarkConfig
 
 class Benchmark(object):
     """Benchmark class can be used to evaluate the model performance, with the objective
@@ -67,9 +69,11 @@ class Benchmark(object):
         self.exp_benchmarker.postprocess = nc_postprocess
 
 
-def benchmark(
+def fit(
     model, config=None, b_dataloader=None, b_func=None
 ):
+    if isinstance(config, BenchmarkConfig):
+        config = Config(benchmark=config)
     benchmarker = ExpBenchmark(config)
     benchmarker.model = model
     if b_func is not None:
@@ -78,3 +82,6 @@ def benchmark(
         benchmarker.b_dataloader = b_dataloader
     benchmarker()
     return benchmarker.results
+
+
+benchmark = fit
