+++ neural_compressor/experimental/quantization.py
@@ -390,12 +390,11 @@ class Quantization(Component):
         return None
 
     @q_func.setter
-    @deprecated(version='2.0', reason="please use `train_func` instead")
     def q_func(self, user_q_func):
-        """Training function for Quantization-Aware Training.
+        """Calibrate quantization parameters for Post-training static quantization.
 
            It is optional and only takes effect when user choose
-           "quant_aware_training" approach in yaml.
+           "post_training_static_quant" approach in yaml.
 
         Args:
             user_q_func: This function takes "model" as input parameter
