+++ neural_compressor/model/base_model.py
@@ -42,12 +42,7 @@ class BaseModel:
     def export(
         self,
         save_path: str,
-        input,
-        target_model_type: str = 'ONNX',
-        quant_format: str = 'QDQ',
-        opset_version: int = 14,
-        *args,
-        **kwargs
+        conf,
     ):
         ''' abstract method of model convertion to ONNX'''
         raise NotImplementedError
