+++ neural_compressor/training.py
@@ -16,8 +16,8 @@
 # limitations under the License.
 
 import copy
-from .conf.pythonic_config import Config, DistillationConfig, Options, \
-                                  PruningConfig, QuantizationAwareTrainingConfig
+from .conf.pythonic_config import Config
+from .config import DistillationConfig, PruningConfig, QuantizationAwareTrainingConfig
 from .experimental.distillation import Distillation
 from .experimental.pruning import Pruning
 from .experimental.quantization import Quantization
@@ -54,8 +54,7 @@ class CompressionManager:
         compression_manager.save("path_to_save")
     """
     def __init__(self, component):
-        self.callbacks = \
-            component.components[0] if isinstance(component, Scheduler) else component
+        self.callbacks = self.CallBacks(component)
         self.model = component.model
         try:
             # TODO: export to ONNX model need original fp32 model now, will remove it
@@ -64,6 +63,46 @@ class CompressionManager:
         except Exception as e:  # pragma: no cover
             logger.warning("Fail to deep copy the model due to {}.".format(repr(e)))
             self.fp32_model = None
+
+    class CallBacks:
+        def __init__(self, component):
+            self.callbacks = \
+                component.components[0] if isinstance(component, Scheduler) else component
+
+        def on_train_begin(self, dataloader=None):
+            """ called before the beginning of epochs"""
+            self.callbacks.on_train_begin(dataloader)
+
+        def on_train_end(self):
+            """ called after the end of epochs"""
+            self.callbacks.on_train_end()
+
+        def on_epoch_begin(self, epoch):
+            """ called on the beginning of epochs"""
+            self.callbacks.on_epoch_begin(epoch)
+
+        def on_step_begin(self, batch_id):
+            """ called on the beginning of batches"""
+            self.callbacks.on_step_begin(batch_id)
+
+        def on_after_compute_loss(self, input, student_output, student_loss, teacher_output=None):
+            """ called on the end of loss computation"""
+            return self.callbacks.on_after_compute_loss(
+                input, student_output, student_loss, teacher_output=None
+            )
+
+        def on_before_optimizer_step(self):
+            """ called on the end of backward"""
+            self.callbacks.on_before_optimizer_step()
+
+
+        def on_step_end(self):
+            """ called on the end of batches"""
+            return self.callbacks.on_step_end()
+
+        def on_epoch_end(self):
+            """ called on the end of epochs"""
+            return self.callbacks.on_epoch_end()
 
     def save(self, root=None):
         """Save compressed model.
@@ -101,7 +140,7 @@ class CompressionManager:
             assert False, "Unsupport export for {} model".format(type(self.model))
 
 
-def prepare_compression(model: Callable, confs: Union[Callable, List], options=None, **kwargs):
+def prepare_compression(model: Callable, confs: Union[Callable, List], **kwargs):
     """_summary_
 
     Args:
@@ -135,20 +174,18 @@ def prepare_compression(model: Callable, confs: Union[Callable, List], options=N
             compression_manager.on_train_end()
     """
 
-    if options is None:
-        options = Options()
     if isinstance(confs, List):
         from .experimental.scheduler import Scheduler
         comps = []
         for conf in confs:
             if isinstance(conf, QuantizationAwareTrainingConfig):
-                conf_ = Config(quantization=conf, options=options)
+                conf_ = Config(quantization=conf)
                 com = Quantization(conf_)
             elif isinstance(conf, PruningConfig):
-                conf_ = Config(pruning=conf, options=options)
+                conf_ = Config(pruning=conf)
                 com = Pruning(conf_)
             elif isinstance(conf, DistillationConfig):
-                conf_ = Config(distillation=conf, options=options)
+                conf_ = Config(distillation=conf)
                 com = Distillation(conf_)
                 assert conf.teacher_model is not None, \
                     "Please set teacher_model in DistillationConfig"
@@ -165,13 +202,13 @@ def prepare_compression(model: Callable, confs: Union[Callable, List], options=N
         component = scheduler
     else:
         if isinstance(confs, QuantizationAwareTrainingConfig):
-            conf = Config(quantization=confs, options=options)
+            conf = Config(quantization=confs)
             component = Quantization(conf)
         elif type(confs) == PruningConfig:
-            conf = Config(pruning=confs, options=options)
+            conf = Config(pruning=confs)
             component = Pruning(conf)
         elif type(confs) == DistillationConfig:
-            conf = Config(distillation=confs, options=options)
+            conf = Config(distillation=confs)
             component = Distillation(conf)
             assert confs.teacher_model is not None, \
                     "Please set teacher_model in DistillationConfig"
