+++ test/benchmark/test_benchmark_2.x.py
@@ -9,33 +9,6 @@ import tempfile
 import re
 from neural_compressor.adaptor.tf_utils.util import write_graph
 
-def build_fake_yaml():
-    fake_yaml = '''
-        model:
-          name: fake_yaml
-          framework: tensorflow
-          inputs: x
-          outputs: op_to_store
-        device: cpu
-        evaluation:
-          accuracy:
-            metric:
-              topk: 1
-          performance:
-            warmup: 5
-            iteration: 10
-            configs:
-               cores_per_instance: 4
-               num_of_instance: 2 
-        tuning:
-          accuracy_criterion:
-            relative: 0.01
-        '''
-    y = yaml.load(fake_yaml, Loader=yaml.SafeLoader)
-    with open('fake_yaml.yaml', "w", encoding="utf-8") as f:
-        yaml.dump(y, f)
-    f.close()
-
 
 def build_benchmark():
     seq = '''
@@ -43,12 +16,14 @@ from argparse import ArgumentParser
 arg_parser = ArgumentParser(description='Parse args')
 arg_parser.add_argument('--input_model', dest='input_model', default='input_model', help='input odel')
 args = arg_parser.parse_args()
-import neural_compressor
+from neural_compressor.benchmark import fit
+from neural_compressor.config import BenchmarkConfig
 from neural_compressor.data import DATASETS
 from neural_compressor.experimental import common
 dataset = DATASETS('tensorflow')['dummy']((100, 32, 32, 1), label=True)
 b_dataloader = common.DataLoader(dataset, batch_size=10)
-neural_compressor.benchmark(args.input_model, 'fake_yaml.yaml', b_dataloader=b_dataloader)
+conf = BenchmarkConfig(warmup=5, iteration=10, cores_per_instance=4, num_of_instance=2)
+fit(args.input_model, conf, b_dataloader=b_dataloader)
     '''
 
     seq1 = '''
@@ -56,14 +31,14 @@ from argparse import ArgumentParser
 arg_parser = ArgumentParser(description='Parse args')
 arg_parser.add_argument('--input_model', dest='input_model', default='input_model', help='input odel')
 args = arg_parser.parse_args()
-import neural_compressor
+from neural_compressor.benchmark import fit
+from neural_compressor.config import BenchmarkConfig
 from neural_compressor.data import DATASETS
 dataset = DATASETS('tensorflow')['dummy']((100, 32, 32, 1), label=True)
 from neural_compressor.experimental import common
-from neural_compressor.conf.config import BenchmarkConf
-conf = BenchmarkConf('fake_yaml.yaml')
+conf = BenchmarkConfig(warmup=5, iteration=10, cores_per_instance=4, num_of_instance=2)
 b_dataloader = common.DataLoader(dataset, batch_size=10)
-neural_compressor.benchmark(args.input_model, conf, b_dataloader=b_dataloader)
+fit(args.input_model, conf, b_dataloader=b_dataloader)
     '''
 
     # test normal case
@@ -88,58 +63,17 @@ def build_benchmark2():
         "arg_parser = ArgumentParser(description='Parse args')\n",
         "arg_parser.add_argument('--input_model', dest='input_model', default='input_model', help='input model')\n",
         "args = arg_parser.parse_args()\n",
-        "import neural_compressor\n"
+        "from neural_compressor.benchmark import fit\n"
         "from neural_compressor.data import DATASETS\n",
         "dataset = DATASETS('tensorflow')['dummy']((5, 32, 32, 1), label=True)\n",
 
         "from neural_compressor.experimental import common\n",
         "b_dataloader = common.DataLoader(dataset)\n",
-        "neural_compressor.benchmark(args.input_model, b_dataloader=b_dataloader)\n"
+        "fit(args.input_model, b_dataloader=b_dataloader)\n"
     ]
-
-    seq1 = '''
-from argparse import ArgumentParser
-arg_parser = ArgumentParser(description='Parse args')
-arg_parser.add_argument('--input_model', dest='input_model', default='input_model', help='input odel')
-args = arg_parser.parse_args()
-import neural_compressor
-from neural_compressor import conf
-from neural_compressor.experimental import common
-conf.evaluation.performance.dataloader.dataset = {'dummy': {'shape': [100,32,32,1], 'label':True}}
-neural_compressor.benchmark(args.input_model, conf)
-    '''
-
-    seq2 = '''
-from argparse import ArgumentParser
-arg_parser = ArgumentParser(description='Parse args')
-arg_parser.add_argument('--input_model', dest='input_model', default='input_model', help='input model')
-args = arg_parser.parse_args()
-
-class Metric:
-    def update(self, pred, label):
-        pass
-
-    def reset(self):
-        pass
-
-    def result(self):
-        return 1.
-
-from neural_compressor import conf
-from neural_compressor.experimental import Benchmark, common
-conf.evaluation.accuracy.dataloader.dataset = {'dummy': {'shape': [100,32,32,1], 'label':True}}
-benchmarker = Benchmark(conf)
-benchmarker.model = args.input_model
-benchmarker.metric = Metric()
-benchmarker.fit('accuracy')
-    '''
 
     with open('fake2.py', "w", encoding="utf-8") as f:
         f.writelines(seq)
-    with open('fake3.py', "w", encoding="utf-8") as f:
-        f.writelines(seq1)
-    with open('fake4.py', "w", encoding="utf-8") as f:
-        f.writelines(seq2)
 
 
 def build_fake_model():
@@ -184,23 +118,16 @@ class TestObjective(unittest.TestCase):
     @classmethod
     def setUpClass(self):
         self.graph_path = build_fake_model()
-        build_fake_yaml()
         build_benchmark()
         build_benchmark2()
         self.cpu_counts = psutil.cpu_count(logical=False)
 
     @classmethod
     def tearDownClass(self):
-        if os.path.exists('fake_yaml.yaml'):
-            os.remove('fake_yaml.yaml')
         if os.path.exists('fake.py'):
             os.remove('fake.py')
-        if os.path.exists('fake.py'):
+        if os.path.exists('fake2.py'):
             os.remove('fake2.py')
-        if os.path.exists('fake.py'):
-            os.remove('fake3.py')
-        if os.path.exists('fake.py'):
-            os.remove('fake4.py')
         if os.path.exists('fake_data_5.py'):
             os.remove('fake_data_5.py')
         if os.path.exists('fake_data_15.py'):
@@ -244,29 +171,6 @@ class TestObjective(unittest.TestCase):
                 self.assertIsNotNone(throughput)
         os.system("rm *.log")
 
-    def test_benchmark_without_yaml(self):
-        os.system("python fake2.py --input_model={} 2>&1 | tee benchmark.log".format(self.graph_path))
-        with open('benchmark.log', "r") as f:
-            for line in f:
-                accuracy = re.search(r"Accuracy is\s+(\d+(\.\d+)?)", line)
-            self.assertIsNotNone(accuracy)
-        os.system("rm *.log")
 
-    def test_benchmark_with_conf(self):
-        os.system("python fake3.py --input_model={}".format(self.graph_path))
-        with open(f'1_{self.cpu_counts}_0.log', "r") as f:
-            for line in f:
-                throughput = re.search(r"Throughput:\s+(\d+(\.\d+)?) images/sec", line)
-            self.assertIsNotNone(throughput)
-        os.system("rm *.log")
- 
-    def test_benchmark_with_custom_metric(self):
-        os.system("python fake4.py --input_model={} 2>&1 | tee benchmark.log".format(self.graph_path))
-        with open('benchmark.log', "r") as f:
-            for line in f:
-                accuracy = re.search(r"Accuracy is\s+(\d+(\.\d+)?)", line)
-            self.assertIsNotNone(accuracy)
-        os.system("rm *.log")
- 
 if __name__ == "__main__":
     unittest.main()
